requirements.txt
flask>=3.0.0
tensorflow>=2.16.0
opencv-python>=4.8.0
numpy>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0
werkzeug>=3.0.0
